+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next version
+
+### 🔧 Fixed
+
+* Fixed FVC `apply_corrections()` only setting the values of robots with invalid transformations.
+
+
 ## 1.7.1 - May 8, 2024
 
 ### ✨ Improved